moved fragments to types folder
