+++ console/console-init/ui/src/graphql-module/queries/endpoint.ts
