+++ console/console-init/ui/src/graphql-module/queries/iot_project.ts
@@ -8,71 +8,12 @@ import gql from "graphql-tag";
 const FILTER_RETURN_IOT_PROJECTS = (filterObject?: any) => {
   const { projectname } = filterObject;
   let filter: string = "";
-
   if (projectname && projectname.trim() !== "") {
     filter += "`$.metadata.name` = '" + projectname + "'";
   }
 
   // TODO: Filters to be incrementally added
   return filter;
-};
-
-const RETURN_ALL_PROJECTS = (filterObj?: any, queryResolver?: string) => {
-  // TODO: Default resolver is subjected to change, with respect to most used query
-  const defaultQueryResolver = `
-    total
-    objects{
-      ... on AddressSpace_consoleapi_enmasse_io_v1beta1 {
-        kind
-        metadata {
-          name
-          namespace
-          creationTimestamp
-        }
-        messagingStatus: status {
-          isReady 
-          phase
-          messages
-        }
-      }
-      ... on IoTProject_iot_enmasse_io_v1alpha1 {
-        kind
-        metadata {
-          name
-          namespace
-          creationTimestamp
-        }
-        iotStatus: status{
-          phase
-          phaseReason 
-        }
-        enabled
-      }
-    }
-  `;
-
-  if (!queryResolver) {
-    queryResolver = defaultQueryResolver;
-  }
-
-  let filter: string = "";
-
-  if (filterObj) {
-    //TODO
-    // filter = FILTER_RETURN_IOT_PROJECTS(filterObj);
-  }
-
-  const IOT_PROJECT_DETAIL = gql(
-    `query allProjects {
-      allProjects(
-        filter: "${filter}",
-      ) {
-         ${queryResolver}
-      }
-    }`
-  );
-
-  return IOT_PROJECT_DETAIL;
 };
 
 const RETURN_IOT_PROJECTS = (filterObj?: any, queryResolver?: string) => {
@@ -100,11 +41,11 @@ const RETURN_IOT_PROJECTS = (filterObj?: any, queryResolver?: string) => {
     queryResolver = defaultQueryResolver;
   }
 
-  let filter: string = "";
+  let filter: string = "`$.kind`='IoTProject'";
 
   if (filterObj) {
-    //TODO
-    FILTER_RETURN_IOT_PROJECTS(filterObj);
+    filter += " AND ";
+    filter += FILTER_RETURN_IOT_PROJECTS(filterObj);
   }
 
   const IOT_PROJECT_DETAIL = gql(
@@ -143,9 +84,4 @@ const TOGGLE_IOT_PROJECTS_STATUS = gql(
   }`
 );
 
-export {
-  RETURN_IOT_PROJECTS,
-  DELETE_IOT_PROJECT,
-  TOGGLE_IOT_PROJECTS_STATUS,
-  RETURN_ALL_PROJECTS
-};
+export { RETURN_IOT_PROJECTS, DELETE_IOT_PROJECT, TOGGLE_IOT_PROJECTS_STATUS };
