+++ console/console-init/ui/src/types/framgmentTypes.json
